Minor separation of concern improvements
